refactor: change expressRouteAdapter folder to main/adapters
