+++ src/main/routes/survey-routes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import { adaptRoute } from '../adapters/express/express-route-adapter'
+import { adaptRoute } from '../adapters/express-route-adapter'
 import { makeAddSurveyController } from '../factories/controllers/add-survey/add-survey-controller-factory'
 
 export default (router: Router): void => {
